backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
sqlalchemy
geoalchemy2
psycopg2-binary
pandas
shapely
requests
python-dotenv
openpyxl
numpy
python-multipart